Rename d3d.h to d3d_utils.h
This renames `d3d.h` to avoid conflicts with the Windows SDK header with the same name.
