+++ foo_ui_columns/d3d_utils.h
